refactor: move GameInstallInfo to its own file
GameInstallInfo is the only type in RomMRomLocal.cs that needs RomM.Settings (it
carries an EmulatorMapping). Splitting it out leaves RomMRomLocal.cs (MainSibling,
RomMRevision, RomMRomLocal) dependency-free so those types can be linked into the
test project. No behaviour change.
